requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
click==8.1.7
clickhouse-connect==0.6.19
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.5.14
Deprecated==1.2.14
distro==1.8.0
duckdb==0.9.1
fastapi==0.85.1
filelock==3.13.1
Flask==3.0.0
Flask-HTTPAuth==4.8.0
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.10.0
funcy==2.0
googleapis-common-protos==1.61.0
google-auth==2.23.4
greenlet==3.0.1
grpcio==1.59.2
grpcio-tools==1.59.2
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hnswlib==0.7.0
hpack==4.0.0
httpcore==1.0.1
httptools==0.6.1
httpx==0.25.1
huggingface-hub==0.17.3
humanfriendly==10.0
hyperframe==6.0.1
charset-normalizer==3.3.2
chromadb==0.3.29
chroma-hnswlib==0.7.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
kubernetes==28.1.0
langchain==0.1.4
langchain-community==0.0.16
langchain-core==0.1.17
langchain-openai==0.0.5
langsmith==0.0.85
lark==1.1.8
load-dotenv==0.1.0
lz4==4.3.2
MarkupSafe==2.1.3
marshmallow==3.20.1
mmh3==4.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.7
numpy==1.26.1
oauthlib==3.2.2
onnxruntime==1.16.1
openai==1.10.0
openapi-schema-pydantic==1.2.4
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-asgi==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
opentelemetry-util-http==0.42b0
overrides==7.4.0
packaging==23.2
pandas==2.1.2
pip==23.3.2
portalocker==2.8.2
posthog==3.0.2
protobuf==4.25.0
psycopg2==2.9.9
pulsar-client==3.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic_core==2.10.1
pydantic==1.10.13
PyPika==0.48.9
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
qdrant-client==1.7.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
setuptools==65.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.20.4
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.2
tokenizers==0.14.1
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
typing-inspect==0.9.0
tzdata==2023.3
urllib3==1.26.18
uvicorn==0.24.0.post1
watchfiles==0.21.0
websocket-client==1.6.4
websockets==12.0
Werkzeug==3.0.1
wrapt==1.15.0
yarl==1.9.2
zipp==3.17.0
zstandard==0.22.0
